Obtain gameid for next / current game for game thread use.
From the schedule page, extract the ESPN gameid for the current / next
game. This ID will be used to generate the content for the Gameday
thread.

I also added a function for determining how long to update the Gameday
thread. It currently gets made four hours before the game and updates
for four hours after the game starts.